requirements-smartsheet.txt
playwright==1.44.0
pymongo==4.7.2
python-dotenv==1.0.1
pydantic==2.5.0
tenacity==8.2.3
prometheus-client==0.19.0
redis==5.0.1
fastapi==0.108.0
uvicorn==0.25.0

sentry-sdk==1.39.0
schedule==1.2.0